refactor: 유틸 함수들 하나의 파일에 정의
