+++ hooks/src/lib/hooks/useCvcValidation.ts
@@ -1,9 +1,7 @@
 import { useEffect, useState } from "react";
 import { ValidationType } from "../../types/validation";
 import { ERROR_MESSAGE } from "../constants/error";
-import isLengthEqual from "../utils/isLengthEqual";
-import isPositiveInteger from "../utils/isPositiveInteger";
-import isEmpty from "../utils/isEmpty";
+import { isEmpty, isLengthEqual, isPositiveInteger } from "../utils/validation";
 
 const MAX_LENGTH = 3;
