--- hooks/src/lib/utils/isEmpty.ts
@@ -1,5 +1,0 @@
-const isEmpty = (value: string) => {
-  return value === "";
-};
-
-export default isEmpty;
