--- hooks/src/lib/utils/isLengthEqual.ts
@@ -1,5 +1,0 @@
-const isLengthEqual = (value: string, length: number) => {
-  return value.length === length;
-};
-
-export default isLengthEqual;
