--- hooks/src/lib/utils/isPositiveInteger.ts
@@ -1,5 +1,0 @@
-const isPositiveInteger = (value: string) => {
-  return /^\d*$/.test(value);
-};
-
-export default isPositiveInteger;
