+++ hooks/src/lib/utils/validation.ts
@@ -1,0 +1,11 @@
+export const isEmpty = (value: string) => {
+  return value === "";
+};
+
+export const isPositiveInteger = (value: string) => {
+  return /^\d*$/.test(value);
+};
+
+export const isLengthEqual = (value: string, length: number) => {
+  return value.length === length;
+};
